bgfinalend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
celery==5.4.0
redis==5.2.0
requests==2.32.3
sentry-sdk[fastapi]==2.18.0
deep-translator==1.11.4
appwrite==17.0.0
minio==7.2.10
qdrant-client==1.12.1
python-dotenv==1.0.1
numpy==2.1.3
opencv-python==4.10.0.84
scikit-learn==1.5.2
Pillow==11.0.0
sentence-transformers==3.3.1
transformers==4.46.3
torchvision==0.20.1
kornia==0.7.4
timm==1.0.11
huggingface-hub==0.26.2
